+++ src/App/index.js
@@ -7,7 +7,7 @@ import SignUpPage from '../pages/RegisterPage/signUpPage';
 import SignInPage from '../pages/RegisterPage/signInPage';
 import PasswordForgetPage from '../pages/RegisterPage/passwordForgetPage.js.js';
 import HomePage from '../pages/HomePage';
-import AccountPage from '../components/Account';
+import AccountPage from '../pages/AccountPage';
 import InfoPage from '../pages/InfoPage';
 import Header, {AccountHeader, ChangeHeader} from '../components/Headers';
 import CompetePage from '../pages/CompetePage';
