+++ src/components/DormSelect/index.js
@@ -8,7 +8,7 @@ import Select from '@material-ui/core/Select';
 
 import { AuthUserContext, withAuthorization } from "../../services/Session";
 import {updateUserDorm, getDorm} from "../../services/Firebase";
-import { assignRanking } from "../../pages/CompetePage/leaderboard.js";
+import { assignRanking } from "../../pages/CompetePage/leaderboard";
 
 const useStyles = makeStyles((theme) => ({
     formControl: {
